requirements.txt
streamlit
pandas
numpy
python-dotenv
langchain-core
langchain-groq
nltk
langchain
huggingface_hub
tqdm